Делает разделение на модули
